added data layer

+++ app/src/main/java/com/sps/todoitems/data/TodoApiModel.kt
@@ -1,3 +1,3 @@
 package com.sps.todoitems.data
 
-data class TodoApiModel(val text: String, val id: Int = -1, val timeStamp: Long)
+data class TodoApiModel(val text: String, val id: Long = -1, val timeStamp: Long)
